backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv>=1.0.1
pydantic>=2.6.4
email-validator>=2.2.0
python-multipart>=0.0.9
firebase-admin>=6.5.0
google-cloud-firestore>=2.16.0
emergentintegrations==0.1.0
google-genai>=0.7.0
Pillow>=10.0.0
httpx>=0.27.0
resend>=2.0.0
twilio>=9.0.0
pytest>=8.0.0
